fix bug requiring exact number of inputs (#2)
Benchmark required the number of fields nf to equal
the number of inputs ni + 1 (for the label). This meant
cases defining expected outputs could not be used.
The new condition is nf >= ni + 1, so now one set of cases
can be used for tests and benchmarks.

Also added some comments to helper functions to add clarity.